Make icons play nice with our Dark UI
Now when you flip to Dark UI the icons won't be black but pixel blue.
Looks normal in the light/normal UI. In fact, I think it looks better
in the light/normal UI because they now match the dashboard icons.

- Remove light statusbar from styles. We're already inheriting the styles from
Settings. Keeping it causes the statusbar to stay dark while using the dark UI.

Change-Id: I8231f62f259d78912547a30b57fd0da3beec2adf